Fix Spring Boot 4 web server factory package relocation
Spring Boot 4 reorganized web server factory classes into dedicated modules:
- ConfigurableServletWebServerFactory: org.springframework.boot.web.servlet.server → org.springframework.boot.web.server.servlet
- TomcatServletWebServerFactory: org.springframework.boot.web.embedded.tomcat → org.springframework.boot.tomcat.servlet
- ErrorPage: org.springframework.boot.web.server → org.springframework.boot.web.error
- addAdditionalTomcatConnectors() → addAdditionalConnectors()

Updated production code and tests in uaa module.
